Extract AbstractHandle from DeliveryHandle
